Correct codesmells

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/time/MasterClock.java
@@ -509,7 +509,7 @@ public class MasterClock implements Serializable {
 				long earthMillisec = (long)(lastPulseTime * MILLISECONDS_PER_MILLISOL);
 
 				// Calculate the actual rate for feedback
-				actualTR = earthMillisec / realElapsedMillisec;
+				actualTR = (double)(earthMillisec / realElapsedMillisec);
 
 				if (!listenerExecutor.isTerminated()
 					&& !listenerExecutor.isShutdown()) {
